+++ packages/ui/Touchable/ClickAllow.js
@@ -1,3 +1,21 @@
+// It's possible that emitting `onTap` will cause a DOM change before the mousedown fires resulting
+// in multiple tap/click events for the same user action. To avoid this, we store the last touchend
+// target and timestamp to compare against the next mouse down. If the timestamp is different (e.g
+// we're on a hybrid device that emitted a touch event but the next was a mouse event) or the target
+// is the same (or no previous target was set if no touch events have been emitted), we allow the
+// mousedown *across Touchable instances*.
+let _lastTouchEnd = {
+	target: null,
+	timeStamp: 0
+};
+
+const shouldAllowMouseDown = (ev) => {
+	return ev.timeStamp !== _lastTouchEnd.timeStamp || (
+		ev.target === _lastTouchEnd.target ||
+		_lastTouchEnd.target === null
+	);
+};
+
 class ClickAllow {
 	constructor () {
 		this.lastTouchEndTime = 0;
@@ -7,6 +25,8 @@ class ClickAllow {
 	setLastTouchEnd (ev) {
 		if (ev && ev.type === 'touchend') {
 			this.lastTouchEndTime = ev.timeStamp;
+			_lastTouchEnd.timeStamp = ev.timeStamp;
+			_lastTouchEnd.target = ev.target;
 		}
 	}
 
@@ -19,7 +39,7 @@ class ClickAllow {
 	shouldAllowMouseEvent (ev) {
 		const {timeStamp} = ev;
 
-		return this.lastTouchEndTime !== timeStamp;
+		return this.lastTouchEndTime !== timeStamp && shouldAllowMouseDown(ev);
 	}
 
 	shouldAllowTap (ev) {